services/ml-service/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
scikit-learn==1.3.2
numpy==1.26.0
kafka-python==2.0.2
joblib==1.3.2
python-dotenv==1.0.0
